Added option for custom degrees
I added a way to add custom degrees for some nodes. So if you want a menu that has unevenly distributed items, you can do that. So if you had 3 radial non button nodes, and you wanted one to be 100 degrees apart from the next and not 120, you could make it 100, 130, 130 degrees apart from each other

To do so I added an optional parameter for degrees you can declare by making a vector full of "degrees." I added a check to make sure degrees is declared, and played with the math a little to make sure that all the nodes that don't have a custom degree are distributed equally around the circle. 

Any undefined degrees ie the degree vector is shorter than item count, the remaining degrees will count as 0. 

If the user's degrees add up to over 360 or whatever max degree they specify, degrees will all reset to 0. 

If the user uses custom degrees for ALL nodes but it adds up to less than 360, then this will add extra angle to compensate. 

If you have any questions feel free to ask. :)